MongoDB 적용

+++ controller/post.mjs
@@ -16,7 +16,7 @@ export async function getPosts(req, res, next) {
 // id를 받아 하나의 포스트를 가져오는 함수
 export async function getPostId(req, res, next) {
   const id = req.params.id;
-  const data = await postRepository.getAllById(id);
+  const data = await postRepository.getById(id);
   if (data) {
     res.status(200).json(data);
   } else {
@@ -26,8 +26,8 @@ export async function getPostId(req, res, next) {
 
 // 포스트를 생성하는 함수
 export async function createPost(req, res, next) {
-  const { userid, name, text } = req.body;
-  const posts = await postRepository.create(userid, name, text);
+  const { text } = req.body;
+  const posts = await postRepository.create(text,req.id);
   res.status(201).json(posts);
 }
 
@@ -35,18 +35,27 @@ export async function createPost(req, res, next) {
 export async function updatePost(req, res, next) {
   const id = req.params.id;
   const text = req.body.text;
-  const post = await postRepository.update(id, text);
-  if (!post) {
-    res.status(404).json({ message: `${id}번의 포스트를 찾을 수 없습니다.` });
-  } else {
-    res.status(200).json(post);
+  const post = await postRepository.getById(id);
+  if(!post){
+    return res.status(404).json({message: `${id}의 포스트가 없습니다.`})
   }
-  next();
-} 
+  if(post.useridx !== req.id){
+    return res.sendStatus(403)
+  }
+    const updated = await postRepository.update(id,text)
+  res.status(200).json(updated)
+  } 
 
 // 포스트 삭제하는 함수
 export async function deletePost(req, res, next) {
   const id = req.params.id;
+  const post = await postRepository.getById(id);
+  if(!post){
+    return res.status(404).json({message: `${id}의 포스트가 없습니다.`})
+  }
+  if(post.useridx !== req.id){
+    return res.sendStatus(403)
+  }
   const posts = await postRepository.remove(id);
   res.status(200).json(posts);
 }
